added changes from gary-dev
