+++ src/gary.js
@@ -1,6 +1,7 @@
 const http = require("axios");
 const auth = require("../auth.json");
 const lib  = require("./lib.js");
+const queries  = require("./queries.js");
 
 const rego   = require("./regocrawl.js");
 
@@ -33,7 +34,6 @@ exports.feats = [
     func: () => Promise.resolve({ file: "https://i.imgur.com/cqJ3cge.gif" })
   },
 
-
   { cmd:  "who's the man",
     log:  () => "asking who the man is.",
     func: (p, user) => Promise.resolve((lib.isTheMan(user)) ? "You da man!" : "You are not the man.")
@@ -41,61 +41,55 @@ exports.feats = [
 
   { cmd:  "tell me a joke",
     log:  () => "asking for a tasteless joke.",
-    func: () => {
-      return http.get("http://api.yomomma.info")
-      .then(response => response.data.joke)
-      .catch(console.error)
-    }
+    func: () =>
+      http.get("http://api.yomomma.info")
+        .then(response => response.data.joke)
+        .catch(console.error)
   },
 
-  { cmd:  "show me a pup.",
+  { cmd:  "show me a pup",
     log:  () => "asking for a pup.",
-    func: () => {
-      return http.get("https://random.dog/woof.json")
-      .then(response => response.data.url)
-      .catch(console.error)
-    }
+    func: () =>
+      http.get("https://random.dog/woof.json")
+        .then(response => response.data.url)
+        .catch(console.error)
   },
 
   { cmd:  "where are you",
     log:  () => "asking for GaryBot's location.",
-    func: () => lib.search("randomPlace")
+    func: () => queries.randomPlace()
   },
 
   { cmd:  "show me some",
     log:  (query) => "searching for some " + query,
-    func: (query) => {
-      return (query != "")
-      ? lib.search("google", query)
-      : Promise.resolve("Show me some what you fuck.")
-    }
+    func: (query) =>
+      (query != "")
+        ? queries.google(query)
+        : Promise.resolve("Show me some what you fuck.")
   },
 
   { cmd:  "simpsons me",
     log:  (query) => "searching for a Simpsons reference - " + query,
-    func: (query) => {
-      return (query != "")
-      ? lib.search("frinkiac", query)
-      : Promise.resolve("Simpsons me what you fuck.")
-    }
+    func: (query) =>
+      (query != "")
+        ? queries.frinkiac(query)
+        : Promise.resolve("Simpsons me what you fuck.")
   },
 
   { cmd:  "tl;dr",
     log:  (query) => "summarising " + query,
-    func: (query) => {
-      return (query != "")
-      ? lib.search("tldr", query)
-      : Promise.resolve("TL;DR what you fuck")
-    }
+    func: (query) =>
+      (query != "")
+        ? queries.frinkiac(query)
+        : Promise.resolve("TL;DR what you fuck")
   },
 
   { cmd:  "what's good on",
     log:  (query) => "looking for popular threads on " + query,
-    func: (query) => {
-      return (query != "")
-        ? lib.search("4chan", query)
+    func: (query) =>
+      (query != "")
+        ? queries.chan(query)
         : Promise.resolve("What's good on what you fuck.")
-    }
   },
 
   { cmd:  "is the following vehicle registered in qld",
